requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.72
botocore==1.35.72
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.15
django-apscheduler==0.7.0
django-debug-toolbar==4.4.6
django-prometheus==2.3.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
et_xmlfile==2.0.0
fontawesomefree==6.6.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prometheus_client==0.21.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyngrok==7.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
s3transfer==0.10.4
sentry-sdk==2.25.1
six==1.16.0
sqlparse==0.5.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.8.2
pyngrok==7.2.3